Rewrote I2c to channel framework

+++ devices/src/main/java/net/audumla/devices/lcd/rpi/RPII2CLCD.java
@@ -14,55 +14,55 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class RPII2CLCD implements net.audumla.devices.lcd.LCD {
-    public static final int DEFAULT_ADDRESS = 0x20;
+    public static final byte DEFAULT_ADDRESS = 0x20;
 
-    protected final static int LCD_8BITMODE = 0x10;
-    protected final static int LCD_4BITMODE = 0x00;
-    protected final static int LCD_2LINE = 0x08;
-    protected final static int LCD_1LINE = 0x00;
-    protected final static int LCD_5x10DOTS = 0x04;
-    protected final static int LCD_5x8DOTS = 0x00;
+    protected final static byte LCD_8BITMODE = 0x10;
+    protected final static byte LCD_4BITMODE = 0x00;
+    protected final static byte LCD_2LINE = 0x08;
+    protected final static byte LCD_1LINE = 0x00;
+    protected final static byte LCD_5x10DOTS = 0x04;
+    protected final static byte LCD_5x8DOTS = 0x00;
 
     // flags for display on/off control
-    protected final static int LCD_DISPLAYON = 0x04;
-    protected final static int LCD_DISPLAYOFF = 0x00;
-    protected final static int LCD_CURSORON = 0x02;
-    protected final static int LCD_CURSOROFF = 0x00;
-    protected final static int LCD_BLINKON = 0x01;
-    protected final static int LCD_BLINKOFF = 0x00;
+    protected final static byte LCD_DISPLAYON = 0x04;
+    protected final static byte LCD_DISPLAYOFF = 0x00;
+    protected final static byte LCD_CURSORON = 0x02;
+    protected final static byte LCD_CURSOROFF = 0x00;
+    protected final static byte LCD_BLINKON = 0x01;
+    protected final static byte LCD_BLINKOFF = 0x00;
 
     // commands
-    protected final static int LCD_FUNCTIONSET = 0x20;
-    protected final static int LCD_CLEARDISPLAY = 0x01;
-    protected final static int LCD_RETURNHOME = 0x02;
-    protected final static int LCD_ENTRYMODESET = 0x04;
-    protected final static int LCD_DISPLAYCONTROL = 0x08;
-    protected final static int LCD_CURSORSHIFT = 0x10;
-    protected final static int LCD_SETCGRAMADDR = 0x40;
-    protected final static int LCD_SETDDRAMADDR = 0x80;
+    protected final static byte LCD_FUNCTIONSET = 0x20;
+    protected final static byte LCD_CLEARDISPLAY = 0x01;
+    protected final static byte LCD_RETURNHOME = 0x02;
+    protected final static byte LCD_ENTRYMODESET = 0x04;
+    protected final static byte LCD_DISPLAYCONTROL = 0x08;
+    protected final static byte LCD_CURSORSHIFT = 0x10;
+    protected final static byte LCD_SETCGRAMADDR = 0x40;
+    protected final static byte LCD_SETDDRAMADDR = (byte) 0x80;
 
     // flags for display/cursor shift
-    protected final static int LCD_DISPLAYMOVE = 0x08;
-    protected final static int LCD_CURSORMOVE = 0x00;
-    protected final static int LCD_MOVERIGHT = 0x04;
-    protected final static int LCD_MOVELEFT = 0x00;
+    protected final static byte LCD_DISPLAYMOVE = 0x08;
+    protected final static byte LCD_CURSORMOVE = 0x00;
+    protected final static byte LCD_MOVERIGHT = 0x04;
+    protected final static byte LCD_MOVELEFT = 0x00;
 
     // flags for display entry mode
-    protected final static int LCD_ENTRYRIGHT = 0x00;
-    protected final static int LCD_ENTRYLEFT = 0x02;
-    protected final static int LCD_ENTRYSHIFTINCREMENT = 0x01;
-    protected final static int LCD_ENTRYSHIFTDECREMENT = 0x00;
+    protected final static byte LCD_ENTRYRIGHT = 0x00;
+    protected final static byte LCD_ENTRYLEFT = 0x02;
+    protected final static byte LCD_ENTRYSHIFTINCREMENT = 0x01;
+    protected final static byte LCD_ENTRYSHIFTDECREMENT = 0x00;
 
     // the I/O expander pinout
-    protected final static int LCD_CHARACTER_WRITE = 0x80;
-    protected final static int LCD_COMMAND = 0x00;
-    protected final static int LCD_BACKLIGHT = 0x02;
-    protected final static int LCD_ENABLE_PIN = 0x40;
-    protected final static int LCD_D4_PIN = 0x20;
-    protected final static int LCD_D5_PIN = 0x10;
-    protected final static int LCD_D6_PIN = 0x08;
-    protected final static int LCD_D7_PIN = 0x04;
-    protected final static int LCD_NO_PIN = 0x00;
+    protected final static byte LCD_CHARACTER_WRITE = (byte) 0x80;
+    protected final static byte LCD_COMMAND = 0x00;
+    protected final static byte LCD_BACKLIGHT = 0x02;
+    protected final static byte LCD_ENABLE_PIN = 0x40;
+    protected final static byte LCD_D4_PIN = 0x20;
+    protected final static byte LCD_D5_PIN = 0x10;
+    protected final static byte LCD_D6_PIN = 0x08;
+    protected final static byte LCD_D7_PIN = 0x04;
+    protected final static byte LCD_NO_PIN = 0x00;
     // used to identify each active bit from an 8 bit character
     protected final static int[] LCD_DATA_4BITMASK = {0x80, 0x40, 0x20, 0x10, 0x8, 0x4, 0x2, 0x1};
     // maps matched bits to 4 bit pins
@@ -70,9 +70,9 @@ public class RPII2CLCD implements net.audumla.devices.lcd.LCD {
     private final String name;
 
     protected PortExtender ext;
-    protected int backlightStatus;
-    private int displayControl;
-    private int displayMode;
+    protected byte backlightStatus;
+    private byte displayControl;
+    private byte displayMode;
 
     private static Map<Integer, net.audumla.devices.lcd.LCD> instances = new HashMap<Integer, net.audumla.devices.lcd.LCD>();
     public static Logger logger = Logger.getLogger(RPII2CLCD.class);
@@ -97,18 +97,18 @@ public class RPII2CLCD implements net.audumla.devices.lcd.LCD {
             synchronized (Thread.currentThread()) {
                 // this will get the RPII2CLCD into the write state to start sending commands
                 Thread.sleep(50, 0);
-                ext.commandWrite(PortExtender.MCP23008_IODIR, 0x00); // all pins to outputs
-                command4bits(LCD_D4_PIN | LCD_D5_PIN);
+                ext.commandWrite(PortExtender.MCP23008_IODIR, (byte) 0x00); // all pins to outputs
+                command4bits((byte) (LCD_D4_PIN | LCD_D5_PIN));
                 Thread.sleep(5, 0);
-                command4bits(LCD_D4_PIN | LCD_D5_PIN);
+                command4bits((byte) (LCD_D4_PIN | LCD_D5_PIN));
                 Thread.sleep(5, 0);
-                command4bits(LCD_D4_PIN | LCD_D5_PIN);
+                command4bits((byte) (LCD_D4_PIN | LCD_D5_PIN));
                 Thread.sleep(1, 0);
                 command4bits(LCD_D5_PIN); // set to 4 bit
-                command4bits(LCD_D5_PIN, LCD_D6_PIN | LCD_D7_PIN); // set lines and character mode
+                command4bits(LCD_D5_PIN, (byte) (LCD_D6_PIN | LCD_D7_PIN)); // set lines and character mode
                 command4bits(LCD_NO_PIN, LCD_D7_PIN); // display off
                 command4bits(LCD_NO_PIN, LCD_D4_PIN); // display clear
-                command4bits(LCD_NO_PIN, LCD_D4_PIN | LCD_D5_PIN | LCD_D6_PIN);
+                command4bits(LCD_NO_PIN, (byte) (LCD_D4_PIN | LCD_D5_PIN | LCD_D6_PIN));
 
                 displayControl = LCD_DISPLAYCONTROL | LCD_DISPLAYON | LCD_CURSOROFF | LCD_BLINKOFF;
                 displayMode = LCD_ENTRYMODESET | LCD_ENTRYLEFT | LCD_ENTRYSHIFTDECREMENT;
@@ -126,31 +126,31 @@ public class RPII2CLCD implements net.audumla.devices.lcd.LCD {
         return name;
     }
 
-    protected void command4bits(int... args) throws Exception {
-        for (int v : args) {
+    protected void command4bits(byte... args) throws Exception {
+        for (byte v : args) {
             send(v, LCD_COMMAND);
         }
     }
 
-    protected void command(int... args) throws Exception {
-        for (int v : args) {
+    protected void command(byte... args) throws Exception {
+        for (byte v : args) {
             send4bits(v, LCD_COMMAND);
         }
     }
 
-    protected void send(int value, int mode) throws Exception {
+    protected void send(byte value, byte mode) throws Exception {
         synchronized (Thread.currentThread()) {
             ext.digitalWrite(backlightStatus);
-            ext.digitalWrite(value | backlightStatus | mode);
-            ext.digitalWrite(value | backlightStatus | mode | LCD_ENABLE_PIN);
+            ext.digitalWrite((byte) (value | backlightStatus | mode));
+            ext.digitalWrite((byte) (value | backlightStatus | mode | LCD_ENABLE_PIN));
             Thread.sleep(0, 500);
-            ext.digitalWrite(value | backlightStatus | mode);
+            ext.digitalWrite((byte) (value | backlightStatus | mode));
             Thread.sleep(0, 50000);
         }
     }
 
-    protected void send4bits(int value, int mode) throws Exception {
-        int bitx4 = 0;
+    protected void send4bits(byte value, byte mode) throws Exception {
+        byte bitx4 = 0;
         for (int i = 0; i < LCD_DATA_4BITMASK.length; ++i) {
             if ((LCD_DATA_4BITMASK[i] & value) > 0) {
                 bitx4 |= LCD_DATA_4BITPIN[i];
@@ -163,8 +163,8 @@ public class RPII2CLCD implements net.audumla.devices.lcd.LCD {
         send(bitx4, mode);
     }
 
-    protected void write(int... args) throws Exception {
-        for (int v : args) {
+    protected void write(byte... args) throws Exception {
+        for (byte v : args) {
             send4bits(v, LCD_CHARACTER_WRITE);
         }
     }
@@ -193,7 +193,7 @@ public class RPII2CLCD implements net.audumla.devices.lcd.LCD {
             row = 3; // we count rows starting w/0
         }
 
-        command(LCD_SETDDRAMADDR | (col + row_offsets[row]));
+        command((byte) (LCD_SETDDRAMADDR | (col + row_offsets[row])));
     }
 
     // Turn the display on/off (quickly)
@@ -238,12 +238,12 @@ public class RPII2CLCD implements net.audumla.devices.lcd.LCD {
     // These commands scroll the display without changing the RAM
     @Override
     public void scrollDisplayLeft() throws Exception {
-        command(LCD_CURSORSHIFT | LCD_DISPLAYMOVE | LCD_MOVELEFT);
+        command((byte) (LCD_CURSORSHIFT | LCD_DISPLAYMOVE | LCD_MOVELEFT));
     }
 
     @Override
     public void scrollDisplayRight() throws Exception {
-        command(LCD_CURSORSHIFT | LCD_DISPLAYMOVE | LCD_MOVERIGHT);
+        command((byte) (LCD_CURSORSHIFT | LCD_DISPLAYMOVE | LCD_MOVERIGHT));
     }
 
     // This is for text that flows Left to Right
@@ -276,10 +276,10 @@ public class RPII2CLCD implements net.audumla.devices.lcd.LCD {
 
     // Allows us to fill the first 8 CGRAM locations
     // with custom characters
-    public void createChar(int location, int charmap[]) throws Exception {
+    public void createChar(int location, byte charmap[]) throws Exception {
         location &= 0x7; // we only have 8 locations 0-7
-        command(LCD_SETCGRAMADDR | (location << 3));
-        for (int i = 0; i < 8; i++) {
+        command((byte) (LCD_SETCGRAMADDR | (location << 3)));
+        for (byte i = 0; i < 8; i++) {
             write(charmap[i]);
         }
     }
@@ -326,13 +326,13 @@ public class RPII2CLCD implements net.audumla.devices.lcd.LCD {
             writeDevice = commandDevice.createChannel(new DeviceRegisterAttr(MCP23008_GPIO));
         }
 
-        public void digitalWrite(int d) throws IOException {
+        public void digitalWrite(byte d) throws IOException {
             writeDevice.write(ByteBuffer.allocateDirect(1).put((byte) d));
         }
 
-        public void commandWrite(int reg, int d) throws IOException {
+        public void commandWrite(int reg, byte d) throws IOException {
             ByteBuffer b = ByteBuffer.allocateDirect(1);
-            commandDevice.setAttribute(b,new DeviceRegisterAttr(reg));
+            commandDevice.setAttribute(b, new DeviceRegisterAttr(reg));
             commandDevice.write(b.put((byte) d));
         }
